Asking for permission when we open the app
